Create `actors_measures` table and `actor_measures` API (#56)
Implements #20  

We need a many to many join table describing an "active" link between an
actor and an activity/measure `actors`, introduced for #18, and
`measures`, from impact (and updated in #31).

##### Permissions by role

- public: none
- registered, no role: none
- analyst: R
- manager: CRUD
- admin: CRUD

##### Rules

- only active actors where allowed for its type (`actor.actortype.is_active = true`)

This also fixes a policy error where we were showing draft records to
anyone, as long as the table had a `draft` column. Now, it will only do
this for people who have an `analyst` role.