Only authorise `base_object` when it exists (#67)
We have an issue where Devise's magic controllers can't run the
`authorize_base_object!` method because they don't have a `base_object`
defined.

The best way to fix this seems to be to only try to run the `authorize`
functionality when `base_object` exists for a controller.

Fixes #66 

While testing this, I've also made an update to `db/seeds.rb` which is
helpful because errors that were nothing to do with the env seeds
method were being suppressed by `rescue NameError`. Instead, we now
check whether the method is defined before calling it so that any errors
can bubble through and we can fix them.