Add Kamal for deployment using containers (#112)
We want to be able to deploy the application to a server which means we
need to configure the server to run the application. We will use Kamal
as our deployment tool so that we can containerise the application.

Given that we don't want to have to store secrets locally, we will use
GitHub Actions to store the secrets and then use them in the deployment
action. These will be triggered when a new release is published, or when
someone with permission triggers the deploy workflow manually.